refactor(Prize): 비즈니스 로직을 가졌으므로, model 레이어로 이동
