+++ src/main/java/lotto/constant/prize/Condition.java
@@ -2,7 +2,7 @@ package lotto.constant.prize;
 
 import java.util.function.BiPredicate;
 
-class Condition {
+public class Condition {
     public static final BiPredicate<Integer, Boolean> FIFTH_PRIZE = (matchedCount, hasBonus) -> matchedCount == 3;
     public static final BiPredicate<Integer, Boolean> FOURTH_PRIZE = (matchedCount, hasBonus) -> matchedCount == 4;
     public static final BiPredicate<Integer, Boolean> THIRD_PRIZE = (matchedCount, hasBonus) -> matchedCount == 5 && !hasBonus;
