+++ src/main/java/lotto/model/WinningNumbers.java
@@ -3,7 +3,6 @@ package lotto.model;
 import java.util.ArrayList;
 import java.util.List;
 import lotto.constant.ExceptionMessage;
-import lotto.constant.prize.Prize;
 
 public record WinningNumbers(Lotto numbers, int bonusNumber) {
     public WinningNumbers {
